Feature #3622: added commitee presence percentage this year
Fixed name of the computed field "meetings_count" to
"meetings_percentage" (to reflect its function).

Added the computed field "meetings_percentage_year" and added an index
on the CommitteeMeeting's date field for it (migration added).

Added a <p> tag to display the date in the committee_member_photos
template - needs translation.